order-service/requirements.txt
Django==4.2.21
djangorestframework==3.15.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
requests==2.32.3
